Include mypy and fix problems
